Updated task.json to enable task usage in DeploymentGroup
Please, update this task (and RunDistributedTests) so it could be used in deployment groups. It would be very useful. The "new" way of running UI tests by using separate agent phase causes many problems when the deployment process is complex. Separating one DeploymentGroup phase into 2 Deployment Group and 1 Agent Phase is sometimes very hard and makes the release more complex (one of the main problems is that the release variables that got their values from the tasks by using logging commands are set back to their original value in the next phase).